MBS-11392 move NoOpTestListener to testFixtures
