+++ subprojects/test-runner/service/src/test/kotlin/com/avito/runner/service/DeviceWorkerTest.kt
@@ -3,6 +3,7 @@ package com.avito.runner.service
 import com.avito.android.Result
 import com.avito.logger.StubLogger
 import com.avito.logger.StubLoggerFactory
+import com.avito.runner.service.listener.NoOpTestListener
 import com.avito.runner.service.model.TestCaseRun
 import com.avito.runner.service.model.intention.State
 import com.avito.runner.service.worker.DeviceWorker
@@ -14,7 +15,6 @@ import com.avito.runner.service.worker.listener.DeviceListener
 import com.avito.runner.service.worker.listener.DeviceLogListener
 import com.avito.runner.service.worker.listener.MessagesDeviceListener
 import com.avito.runner.service.worker.listener.StubDeviceListener
-import com.avito.runner.test.NoOpTestListener
 import com.avito.runner.test.StubActionResult
 import com.avito.runner.test.StubDevice
 import com.avito.runner.test.StubDevice.Companion.installApplicationSuccess
