+++ subprojects/test-runner/service/src/testFixtures/kotlin/com/avito/runner/service/listener/NoOpTestListener.kt
@@ -1,7 +1,6 @@
-package com.avito.runner.test
+package com.avito.runner.service.listener
 
 import com.avito.android.Result
-import com.avito.runner.service.listener.TestListener
 import com.avito.runner.service.model.TestCase
 import com.avito.runner.service.model.TestCaseRun
 import com.avito.runner.service.worker.device.Device
